+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [Unreleased](https://github.com/jshwi/constcheck/compare/v0.4.0...HEAD)
 ------------------------------------------------------------------------
+### Fixed
+- Files can also be used as path arguments
 
 [0.4.0](https://github.com/jshwi/constcheck/releases/tag/v0.4.0) - 2022-04-25
 ------------------------------------------------------------------------